+++ pl_bolts/callbacks/__init__.py
@@ -3,6 +3,7 @@ from pl_bolts.callbacks.byol_updates import BYOLMAWeightUpdate
 from pl_bolts.callbacks.data_monitor import ModuleDataMonitor, TrainingDataMonitor
 from pl_bolts.callbacks.printing import PrintTableMetricsCallback
 from pl_bolts.callbacks.ssl_online import SSLOnlineEvaluator
+from pl_bolts.callbacks.torch_ort import ORTCallback
 from pl_bolts.callbacks.variational import LatentDimInterpolator
 from pl_bolts.callbacks.verification.batch_gradient import BatchGradientVerificationCallback  # type: ignore
 from pl_bolts.callbacks.vision.confused_logit import ConfusedLogitCallback
@@ -18,4 +19,5 @@ __all__ = [
     "LatentDimInterpolator",
     "ConfusedLogitCallback",
     "TensorboardGenerativeModelImageSampler",
+    "ORTCallback",
 ]